move WebappMicroservice class to internal.http.msf4j package
