+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/http/msf4j/WebappMicroservice.java
@@ -16,11 +16,10 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal.http;
+package org.wso2.carbon.uis.internal.http.msf4j;
 
 import org.wso2.carbon.uis.api.http.HttpRequest;
 import org.wso2.carbon.uis.api.http.HttpResponse;
-import org.wso2.carbon.uis.internal.http.msf4j.Msf4jHttpRequest;
 import org.wso2.msf4j.Microservice;
 import org.wso2.msf4j.Request;
